neural net implementation by @dmittov
